RF: Split module into files.
